pyproject.toml
[project]
name = "agentic-rag-gri"
version = "0.1.0"
description = "Système RAG agentique pour le Guide de Référence pour l'Innovation (GRI) des FAR"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Omar Piro" }]
keywords = ["rag", "llm", "gri", "far", "iso-15288", "systems-engineering"]

dependencies = [
    # === Core Framework ===
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # === LLM & Embeddings (Hugging Face) ===
    "huggingface-hub>=0.20.0",
    "sentence-transformers>=2.2.2",
    "transformers>=4.36.0",
    "torch>=2.1.0",

    # === Vector Store ===
    "qdrant-client>=1.7.0",
    "chromadb>=0.4.22",  # dev/fallback

    # === Sparse Search ===
    "rank-bm25>=0.2.2",

    # === Document Processing ===
    "python-docx>=1.1.0",
    "pypandoc>=1.12",

    # === Async & HTTP ===
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "sse-starlette>=2.0.0",

    # === Logging & Config ===
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # === Evaluation ===
    "ragas>=0.1.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",

    # === Utilities ===
    "tenacity>=8.2.3",
    "tiktoken>=0.5.2",

    # === Production API ===
    "slowapi>=0.1.9",
    "prometheus-fastapi-instrumentator>=6.1.0",

    # === State Management ===
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
]

[project.scripts]
gri-ingest = "src.ingestion.pipeline:main"
gri-eval = "src.evaluation.pipeline:main"
gri-agent = "src.agents.orchestrator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# === Ruff (Linting) ===
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends)
    "B904",   # raise without from (intentional in some cases)
]

[tool.ruff.isort]
known-first-party = ["src"]

# === Black (Formatting) ===
[tool.black]
line-length = 100
target-version = ["py311"]

# === Mypy (Type Checking) ===
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "rank_bm25.*",
    "sentence_transformers.*",
    "qdrant_client.*",
    "chromadb.*",
    "ragas.*",
    "docx.*",
    "asyncpg.*",
    "redis.*",
]
ignore_missing_imports = true

# === Pytest ===
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# === Coverage ===
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
